treewide: format all inactive Nix files
After final improvements to the official formatter implementation,
this commit now performs the first treewide reformat of Nix files using it.
This is part of the implementation of RFC 166.

Only "inactive" files are reformatted, meaning only files that
aren't being touched by any PR with activity in the past 2 months.
This is to avoid conflicts for PRs that might soon be merged.
Later we can do a full treewide reformat to get the rest,
which should not cause as many conflicts.

A CI check has already been running for some time to ensure that new and
already-formatted files are formatted, so the files being reformatted here
should also stay formatted.

This commit was automatically created and can be verified using

    nix-build https://github.com/infinisil/treewide-nixpkgs-reformat-script/archive/a08b3a4d199c6124ac5b36a889d9099b4383463f.tar.gz \
      --argstr baseRev f84e54abd39b19566686e3201627f618736f06b4
    result/bin/apply-formatting $NIXPKGS_PATH

+++ nixos/modules/services/web-apps/sogo.nix
@@ -1,4 +1,11 @@
-{ config, pkgs, lib, ... }: with lib; let
+{
+  config,
+  pkgs,
+  lib,
+  ...
+}:
+with lib;
+let
   cfg = config.services.sogo;
 
   preStart = pkgs.writeShellScriptBin "sogo-prestart" ''
@@ -6,17 +13,29 @@
     chown sogo:sogo /etc/sogo/sogo.conf
     chmod 640 /etc/sogo/sogo.conf
 
-    ${if (cfg.configReplaces != {}) then ''
-      # Insert secrets
-      ${concatStringsSep "\n" (mapAttrsToList (k: v: ''export ${k}="$(cat "${v}" | tr -d '\n')"'') cfg.configReplaces)}
+    ${
+      if (cfg.configReplaces != { }) then
+        ''
+          # Insert secrets
+          ${concatStringsSep "\n" (
+            mapAttrsToList (k: v: ''export ${k}="$(cat "${v}" | tr -d '\n')"'') cfg.configReplaces
+          )}
 
-      ${pkgs.perl}/bin/perl -p ${concatStringsSep " " (mapAttrsToList (k: v: '' -e 's/${k}/''${ENV{"${k}"}}/g;' '') cfg.configReplaces)} /etc/sogo/sogo.conf.raw > /etc/sogo/sogo.conf
-    '' else ''
-      cp /etc/sogo/sogo.conf.raw /etc/sogo/sogo.conf
-    ''}
+          ${pkgs.perl}/bin/perl -p ${
+            concatStringsSep " " (
+              mapAttrsToList (k: v: ''-e 's/${k}/''${ENV{"${k}"}}/g;' '') cfg.configReplaces
+            )
+          } /etc/sogo/sogo.conf.raw > /etc/sogo/sogo.conf
+        ''
+      else
+        ''
+          cp /etc/sogo/sogo.conf.raw /etc/sogo/sogo.conf
+        ''
+    }
   '';
 
-in {
+in
+{
   options.services.sogo = with types; {
     enable = mkEnableOption "SOGo groupware";
 
@@ -53,7 +72,7 @@ in {
         specified file.
       '';
       type = attrsOf str;
-      default = {};
+      default = { };
       example = {
         LDAP_BINDPW = "/var/lib/secrets/sogo/ldappw";
       };
@@ -88,7 +107,13 @@ in {
 
     systemd.services.sogo = {
       description = "SOGo groupware";
-      after = [ "postgresql.service" "mysql.service" "memcached.service" "openldap.service" "dovecot2.service" ];
+      after = [
+        "postgresql.service"
+        "mysql.service"
+        "memcached.service"
+        "openldap.service"
+        "dovecot2.service"
+      ];
       wantedBy = [ "multi-user.target" ];
       restartTriggers = [ config.environment.etc."sogo/sogo.conf.raw".source ];
 
@@ -169,14 +194,23 @@ in {
     systemd.services.sogo-ealarms = {
       description = "SOGo email alarms";
 
-      after = [ "postgresql.service" "mysqld.service" "memcached.service" "openldap.service" "dovecot2.service" "sogo.service" ];
+      after = [
+        "postgresql.service"
+        "mysqld.service"
+        "memcached.service"
+        "openldap.service"
+        "dovecot2.service"
+        "sogo.service"
+      ];
       restartTriggers = [ config.environment.etc."sogo/sogo.conf.raw".source ];
 
       startAt = [ "minutely" ];
 
       serviceConfig = {
         Type = "oneshot";
-        ExecStart = "${pkgs.sogo}/bin/sogo-ealarms-notify${optionalString (cfg.ealarmsCredFile != null) " -p ${cfg.ealarmsCredFile}"}";
+        ExecStart = "${pkgs.sogo}/bin/sogo-ealarms-notify${
+          optionalString (cfg.ealarmsCredFile != null) " -p ${cfg.ealarmsCredFile}"
+        }";
 
         ProtectSystem = "strict";
         ProtectHome = true;
@@ -255,13 +289,14 @@ in {
         alias ${pkgs.sogo}/lib/GNUstep/SOGo/$1.SOGo/Resources/$2;
       '';
 
-      locations."~ ^/SOGo/so/ControlPanel/Products/[^/]*UI/Resources/.*\\.(jpg|png|gif|css|js)$".extraConfig = ''
-        alias ${pkgs.sogo}/lib/GNUstep/SOGo/$1.SOGo/Resources/$2;
-      '';
+      locations."~ ^/SOGo/so/ControlPanel/Products/[^/]*UI/Resources/.*\\.(jpg|png|gif|css|js)$".extraConfig =
+        ''
+          alias ${pkgs.sogo}/lib/GNUstep/SOGo/$1.SOGo/Resources/$2;
+        '';
     };
 
     # User and group
-    users.groups.sogo = {};
+    users.groups.sogo = { };
     users.users.sogo = {
       group = "sogo";
       isSystemUser = true;
